Move headers classes to own package
