+++ tutor/src/main/java/org/tudalgo/algoutils/transform/util/ForceSignatureAnnotationProcessor.java
@@ -2,6 +2,8 @@ package org.tudalgo.algoutils.transform.util;
 
 import org.objectweb.asm.*;
 import org.tudalgo.algoutils.student.annotation.ForceSignature;
+import org.tudalgo.algoutils.transform.util.headers.FieldHeader;
+import org.tudalgo.algoutils.transform.util.headers.MethodHeader;
 
 import java.util.ArrayList;
 import java.util.HashMap;
