+++ tutor/src/main/java/org/tudalgo/algoutils/transform/util/headers/FieldHeader.java
@@ -1,9 +1,11 @@
-package org.tudalgo.algoutils.transform.util;
+package org.tudalgo.algoutils.transform.util.headers;
 
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.FieldVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
+import org.tudalgo.algoutils.transform.util.Constants;
+import org.tudalgo.algoutils.transform.util.TransformationUtils;
 
 import java.lang.reflect.Field;
 import java.util.Objects;
