+++ tutor/src/main/java/org/tudalgo/algoutils/transform/util/headers/Header.java
@@ -1,7 +1,9 @@
-package org.tudalgo.algoutils.transform.util;
+package org.tudalgo.algoutils.transform.util.headers;
 
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
+import org.tudalgo.algoutils.transform.util.Constants;
+import org.tudalgo.algoutils.transform.util.TransformationUtils;
 
 import java.util.Arrays;
