+++ tutor/src/main/java/org/tudalgo/algoutils/transform/util/headers/MethodHeader.java
@@ -1,8 +1,10 @@
-package org.tudalgo.algoutils.transform.util;
+package org.tudalgo.algoutils.transform.util.headers;
 
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
+import org.tudalgo.algoutils.transform.util.Constants;
+import org.tudalgo.algoutils.transform.util.TransformationUtils;
 
 import java.lang.reflect.Executable;
 import java.lang.reflect.Method;
